[SNAPPYDATA][MERGE-2.1] Missing patches and version changes
- updated optimized ByteArrayMethods.arrayEquals as per the code in Spark 2.1
  - adapt the word alignment code and optimize it a bit
  - in micro-benchmarks the new method is 30-60% faster than upstream version;
    at larger sizes it is 40-50% faster meaning its base word comparison loop itself is faster
- increase default locality time from 3s to 10s since the previous code to force
  executor-specific routing if it is alive has been removed
- added back cast removal optimization when types differ only in nullability
- add serialization and proper nanoTime handling from *CpuTime added in Spark 2.1.x;
  use DoubleAccumulator for these new fields like done for others to get more accurate results;
  also avoid the rare conditions where these cpu times could be negative
- cleanup handling of jobId and related new fields in Task with kryo serialization
- reverted change to AnalysisException with null check for plan since it is transient now
- reverted old Spark 2.0 code that was retained in InsertIntoTable and changed to Spark 2.1 code
- updated library versions and make them uniform as per upstream Spark for
  commons-lang3, metrics-core, py4j, breeze, univocity; also updated exclusions as
  per the changes to Spark side between 2.0.2 to 2.1.0
- added gradle build for the new mesos sub-project

+++ core/src/main/scala/org/apache/spark/executor/TaskMetrics.scala
@@ -48,9 +48,9 @@ import org.apache.spark.util._
 class TaskMetrics private[spark] () extends Serializable with KryoSerializable {
   // Each metric is internally represented as an accumulator
   private val _executorDeserializeTime = new DoubleAccumulator
-  private val _executorDeserializeCpuTime = new LongAccumulator
+  private val _executorDeserializeCpuTime = new DoubleAccumulator
   private val _executorRunTime = new DoubleAccumulator
-  private val _executorCpuTime = new LongAccumulator
+  private val _executorCpuTime = new DoubleAccumulator
   private val _resultSize = new LongAccumulator
   private val _jvmGCTime = new LongAccumulator
   private val _resultSerializationTime = new DoubleAccumulator
@@ -67,7 +67,7 @@ class TaskMetrics private[spark] () extends Serializable with KryoSerializable {
   /**
    * CPU Time taken on the executor to deserialize this task in nanoseconds.
    */
-  def executorDeserializeCpuTime: Long = _executorDeserializeCpuTime.sum
+  def executorDeserializeCpuTime: Long = _executorDeserializeCpuTime.sum.toLong
 
   /**
    * Time the executor spends actually running the task (including fetching shuffle data).
@@ -78,7 +78,7 @@ class TaskMetrics private[spark] () extends Serializable with KryoSerializable {
    * CPU Time the executor spends actually running the task
    * (including fetching shuffle data) in nanoseconds.
    */
-  def executorCpuTime: Long = _executorCpuTime.sum
+  def executorCpuTime: Long = _executorCpuTime.sum.toLong
 
   /**
    * The number of bytes this task transmitted back to the driver as the TaskResult.
@@ -125,10 +125,10 @@ class TaskMetrics private[spark] () extends Serializable with KryoSerializable {
   // Setters and increment-ers
   private[spark] def setExecutorDeserializeTime(v: Double): Unit =
     _executorDeserializeTime.setValue(v)
-  private[spark] def setExecutorDeserializeCpuTime(v: Long): Unit =
+  private[spark] def setExecutorDeserializeCpuTime(v: Double): Unit =
     _executorDeserializeCpuTime.setValue(v)
   private[spark] def setExecutorRunTime(v: Double): Unit = _executorRunTime.setValue(v)
-  private[spark] def setExecutorCpuTime(v: Long): Unit = _executorCpuTime.setValue(v)
+  private[spark] def setExecutorCpuTime(v: Double): Unit = _executorCpuTime.setValue(v)
   private[spark] def setResultSize(v: Long): Unit = _resultSize.setValue(v)
   private[spark] def setJvmGCTime(v: Long): Unit = _jvmGCTime.setValue(v)
   private[spark] def setResultSerializationTime(v: Double): Unit =
